Simplify: extract shared utils, fix bugs found in code review
- Extract SOURCE_LABELS, ACTION_LABELS, badge variant functions,
  formatTimestamp, and relativeTime into utils/audit-utils.ts to
  eliminate duplication between Browse.tsx and Detail.tsx
- Fix CopyButton setTimeout leak: use useEffect with cleanup to
  prevent state updates on unmounted components
- Fix browseWithFilter passing full ISO strings to datetime-local
  inputs (slice to YYYY-MM-DDTHH:mm format)
- Remove broken status code range filter (backend does exact-match
  on StatusCode, so "4xx" label was misleading — only matched 400)
- Remove unnecessary navigate() wrapper in Dashboard
- Remove redundant setState calls before navigation in applyDatePreset
- Fix pagination ellipsis keys to avoid array-index-as-key lint error
- Add aria-hidden to CopyButton SVGs for accessibility
